Intake functions
- Split full collect into two separate CGs, Herd and Eat
- Added ele pinch toggle for manual use
- Removed some functions
